requirements.txt
numpy==1.24.4
pandas==2.0.3
faiss-cpu==1.8.0
langchain-community==0.4.1